move classes to their package
